examples/dense-schema/pyproject.toml
[project]
name = "dbxcarta-dense-schema-example"
version = "0.1.0"
description = "Dense single-schema fixture for dbxcarta Phase G evaluation (500/1000 tables)"
authors = [{ name = "Ryan Knight" }]
requires-python = ">=3.12"
dependencies = [
    "dbxcarta>=0.2.32",
    "databricks-sdk>=0.40",
    "python-dotenv",
    "orjson",
]

[project.optional-dependencies]
test = ["pytest"]

[project.scripts]
dbxcarta-dense-generate = "dbxcarta_dense_schema_example.generator:main"
dbxcarta-dense-materialize = "dbxcarta_dense_schema_example.materialize:main"
dbxcarta-dense-generate-questions = "dbxcarta_dense_schema_example.question_generator:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
dbxcarta = { path = "../..", editable = true }

[tool.pytest.ini_options]
addopts = "-m 'not live'"
markers = [
    "live: tests that require a live Databricks workspace and Neo4j instance",
]
